refactor: rename dataPipeline.ts to data-pipeline.ts
Match the kebab-case convention used by other files in
`enterprise/services/` (agent-handler.ts, ai-access.ts,
conversation-buffer.ts, etc.).
